bot/requirements.txt
tensorflow
opencv-python
python-telegram-bot==20.3
pillow
